+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 - Partial shuffle mutation operator
+- [STABLE] Uniform crossover operator
 
 ## [1.1.1] - 2023-12-23